requirements.txt
cytoolz==0.11.2
eth-brownie==1.17.2
pytest==6.2.5
#vyper==0.3.1
#vyper @ git+https://github.com/vyperlang/vyper@123fbc332214f93edddf7abb3f052eed47ddfa79
#vyper @ git+https://github.com/ActorForth/vyper@95e318d5f2978de2307607976721b135e9336efa
vyper @ git+https://github.com/ActorForth/vyper@0f9ff286b1ab5d6a0ce91dcd92926cfdb8b96220
web3==5.25.0
pytest-bdd==5.0.0
pytest-asyncio==0.18.1
